Added: CompatService to RhtsOptions
This option defines whether the rhts-compat service keeps running even
when test does not run in compatible mode.

When -CompatService is given, and the task does not run in compat.mode
the service is stopped otherwise it is left as is. The service will be
restarted when it's required by next task.