rhts-reboot can't rely on BootCurrent
Some EFI implementations do not provide the BootCurrent variable, so we
need some other way to know which entry to set BootNext to when
rebooting.

A corresponding Beaker patch writes out /root/EFI_BOOT_ENTRY.TXT after
installation, so that rhts-reboot can use that as a fallback if
BootCurrent is not available.

Bug: 1031876
Change-Id: Ia8228faa8761e663e0687dd2cc43cf12b1bdb514