Change default task tagging, and RPM names.
'make build' will now build an RPM without the repo prepended to it.
Custom virtual provides can also be added.

For 'make tag' to determine what version it needs to use in the tag,
it will first attempt to get the version from a new style tag (no repo
name prepended), and if that fails will attempt to use an old style tag (repo
name prepended).

Bug:1044913
Change-Id: I981aa580f0e1fceba702eaafe2c4281b2a0f72f6