make rhts-lint and rhts-mk-generate-specfile deal in unicode not bytes
This is the missing portion of commit db3954f2, which made the same
change for just the rhts.testinfo module. But these other scripts do
their own file handling and thus need a similar treatment to ensure they
decode/encode as UTF-8 and then pass unicode objects to the
rhts.testinfo module.

Bug: 1491658
Change-Id: I0ad83daa66eb7f7d2ff8de376ec7174783797a5b